refactor(api): move guardrails + replay dashboard endpoints
Commit 4 of the dashboard fold-in. Lifts 6 endpoints out of
dashboard/api/main.py into two new files under
server/api/dashboard/:

- guardrails.py: GET /api/guardrails/events, /api/guardrails/aggregate
                 plus the _guardrail_since + _validate_guardrail_event_filter
                 helpers they share (move; no callers outside this file).
- replay.py:     GET /api/sessions/{id}/replay, DELETE same,
                 GET /api/replay/storage, POST /api/projects/{id}/replay/retention.

Behaviour is preserved verbatim: guardrails handlers keep their
ACTIVE_GUARDRAIL_ACTIONS / GUARDRAIL_CATEGORIES validation + the
HTTPException(400) on unknown category/action, replay endpoints keep
the raw text() select on sessions.replay_size_bytes and the in-memory
project.replay.retention_days mutation.

tests/server/test_replay_endpoint.py repoints at the daemon's
build_app(...) so the moved handlers stay reachable. The guardrails
aggregate test in test_dashboard_api.py already runs through the
daemon's client fixture from commit 2; no extra change needed.

ruff check src + uv run mypy + uv run pytest -q all green:
1666 passed, 4 skipped.
