+++ src/dashboard/api/main.py
@@ -12,17 +12,10 @@ from fastapi import Body, FastAPI, File, HTTPException, Query, UploadFile
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import FileResponse
 from fastapi.staticfiles import StaticFiles
-from sqlalchemy import text
 
 from voicegateway.core.auth import resolve_cors_origins
 from voicegateway.repository import (
-    guardrail_events_repository as guardrail_events,
-)
-from voicegateway.repository import (
     latency_observations_repository as latency_observations,
-)
-from voicegateway.repository import (
-    replay_repository as replay,
 )
 from voicegateway.repository import (
     tenants_repository as tenants,
@@ -109,217 +102,6 @@ def configure(gateway: Gateway) -> None:
         allow_headers=["*"],
     )
     _cors_configured = True
-
-
-@app.get("/api/guardrails/events")
-async def get_guardrail_events(
-    days: int = Query(7, ge=1, le=365),
-    project: str | None = Query(None),
-    tenant: str | None = Query(None),
-    category: str | None = Query(None),
-    action: str | None = Query(None),
-    event_type: str | None = Query(None, pattern="^(fired|bypassed)$"),
-    limit: int = Query(100, ge=1, le=1000),
-) -> dict[str, Any]:
-    """Return filtered guardrail audit events for the dashboard."""
-    gw = _get_gateway()
-    if gw.storage is None:
-        return {"events": [], "filter": {"project": project, "tenant": tenant}}
-    _validate_guardrail_event_filter(category=category, action=action)
-    await gw.storage._ensure_initialized()
-    async with gw.storage._conn.session() as db:
-        rows = await guardrail_events.list_events(
-            db,
-            since=_guardrail_since(days),
-            project=project,
-            tenant=tenant,
-            category=category,
-            action=action,
-            event_type=event_type,
-            limit=limit,
-        )
-    return {
-        "events": [dataclasses.asdict(row) for row in rows],
-        "filter": {
-            "days": days,
-            "project": project,
-            "tenant": tenant,
-            "category": category,
-            "action": action,
-            "event_type": event_type,
-        },
-    }
-
-
-@app.get("/api/guardrails/aggregate")
-async def get_guardrail_aggregate(
-    days: int = Query(7, ge=1, le=365),
-    project: str | None = Query(None),
-    tenant: str | None = Query(None),
-    category: str | None = Query(None),
-) -> dict[str, Any]:
-    """Return guardrail counts plus optional category drilldown."""
-    gw = _get_gateway()
-    if gw.storage is None:
-        return {"counts": [], "top_sessions": []}
-    _validate_guardrail_event_filter(category=category, action=None)
-    await gw.storage._ensure_initialized()
-    async with gw.storage._conn.session() as db:
-        since = _guardrail_since(days)
-        counts = await guardrail_events.aggregate_counts(
-            db, since=since, project=project, tenant=tenant
-        )
-        top_sessions = (
-            await guardrail_events.top_sessions_by_category(
-                db,
-                category=category,
-                since=since,
-                project=project,
-                tenant=tenant,
-            )
-            if category
-            else []
-        )
-    return {
-        "counts": [dataclasses.asdict(row) for row in counts],
-        "top_sessions": [dataclasses.asdict(row) for row in top_sessions],
-        "filter": {
-            "days": days,
-            "project": project,
-            "tenant": tenant,
-            "category": category,
-        },
-    }
-
-
-# ----------------------------------------------------------------------
-# v0.3.0 conversation replay (REQ-VG-REPLAY-001..006).
-#
-# Four handlers backing the dashboard Replay page (T11) and the
-# storage-usage view. Same `/api/*` prefix as the v0.2.0 metrics
-# endpoints, deviating from the Foundry's literal `/v1/*` prefix per
-# the dashboard convention; the main server may publish `/v1/*` mirrors
-# later if SDK consumers need them.
-# ----------------------------------------------------------------------
-
-
-@app.get("/api/sessions/{session_id}/replay")
-async def get_session_replay(session_id: str) -> dict[str, Any]:
-    """Return the full time-ordered replay for one session.
-
-    Used by the Replay page (T11) to pre-fetch the full timeline on
-    page load (OQ3 resolution: pre-fetch over streaming). Each event
-    carries its modality so the consumer routes to the right pane.
-
-    Empty list when the session has no captured replay events
-    (pre-v0.3.0 sessions or projects with ``replay.enabled: false``);
-    the FE renders a PreV030Banner in that case (REQ-VG-REPLAY-001
-    AC-3).
-    """
-    gw = _get_gateway()
-    if gw.storage is None:
-        raise HTTPException(status_code=503, detail="Storage not configured")
-    await gw.storage._ensure_initialized()
-    async with gw.storage._conn.session() as db:
-        events = await replay.read_full_replay(db, session_id)
-    return {
-        "session_id": session_id,
-        "events": [dataclasses.asdict(e) for e in events],
-    }
-
-
-@app.delete("/api/sessions/{session_id}/replay")
-async def delete_session_replay(session_id: str) -> dict[str, Any]:
-    """Delete every replay row tied to a session (REQ-VG-REPLAY-006 AC-3).
-
-    Cascade across all four ``replay_*`` tables in one transaction.
-    Returns the total row count deleted (across modalities) so the
-    caller can confirm the cleanup landed.
-    """
-    gw = _get_gateway()
-    if gw.storage is None:
-        raise HTTPException(status_code=503, detail="Storage not configured")
-    await gw.storage._ensure_initialized()
-    async with gw.storage._conn.session() as db:
-        deleted = await replay.delete_replay(db, session_id)
-    return {"session_id": session_id, "deleted_rows": deleted}
-
-
-@app.get("/api/replay/storage")
-async def get_replay_storage() -> dict[str, Any]:
-    """Return per-project replay storage breakdown.
-
-    Sums ``sessions.replay_size_bytes`` per project (the column is
-    populated by T08's ``finalize_session_replay``). Sessions without
-    captured replay (NULL replay_size_bytes) contribute 0 to the sum.
-    The Refinery requires the dashboard surface this so "the cost is
-    not invisible" to the developer.
-    """
-    gw = _get_gateway()
-    if gw.storage is None:
-        raise HTTPException(status_code=503, detail="Storage not configured")
-    await gw.storage._ensure_initialized()
-    async with gw.storage._conn.session() as db:
-        result = await db.execute(
-            text(
-                "SELECT project, COALESCE(SUM(replay_size_bytes), 0) "
-                "FROM sessions "
-                "WHERE replay_size_bytes IS NOT NULL "
-                "GROUP BY project "
-                "ORDER BY project ASC"
-            )
-        )
-        per_project = []
-        total = 0
-        for row in result:
-            project, size = row
-            size_int = int(size or 0)
-            per_project.append(
-                {
-                    "project": project,
-                    "replay_size_bytes": size_int,
-                }
-            )
-            total += size_int
-        return {
-            "total_replay_size_bytes": total,
-            "by_project": per_project,
-        }
-
-
-@app.post("/api/projects/{project_id}/replay/retention")
-async def update_replay_retention(
-    project_id: str,
-    body: dict[str, Any] = Body(...),
-) -> dict[str, Any]:
-    """Update the retention window for one project.
-
-    Body shape: ``{"retention_days": N}`` (1..365 inclusive). The
-    update is applied in-memory to ``gw.config.projects[project_id]
-    .replay.retention_days`` so the retention worker (T06) picks it
-    up on its next tick.
-
-    v0.3.0 limitation: the update is not persisted to ``voicegw.yaml``
-    on disk. Restart re-reads the original value. Persistence is a
-    follow-up that needs config-file-write infrastructure; the in-
-    memory mutation matches the runtime contract the retention worker
-    reads from.
-    """
-    gw = _get_gateway()
-    new_days_raw = body.get("retention_days")
-    if not isinstance(new_days_raw, int) or new_days_raw < 1 or new_days_raw > 365:
-        raise HTTPException(
-            status_code=422,
-            detail="retention_days must be an int in [1, 365]",
-        )
-    project = gw.config.projects.get(project_id) if gw.config else None
-    if project is None:
-        raise HTTPException(status_code=404, detail=f"Project '{project_id}' not found")
-    project.replay.retention_days = new_days_raw
-    return {
-        "project_id": project_id,
-        "retention_days": new_days_raw,
-    }
 
 
 @app.get("/api/providers/by-project")
